Correzione indirizzi e aggiunta openocd nel Makefile

+++ stm32_startup.c
@@ -175,7 +175,7 @@ void Default_Handler(void){
 void Reset_Handler(void){
 
     //copy .data section to SRAM
-    uint32_t size = &_edata - &_sdata;
+    uint32_t size = (uint32_t)&_edata - (uint32_t)&_sdata;
 
     uint8_t *pDst = (uint8_t *)&_sdata; //sram
     uint8_t *pSrc = (uint8_t *)&_etext; //flash
@@ -184,7 +184,7 @@ void Reset_Handler(void){
                            // Sram all'indirizzo puntato pDst, sposta inoltre tutti i puntatori avanti
                            // di un byte
     }
-    size = &_ebss - &_sbss;
+    size = (uint32_t)(&_ebss) - (uint32_t)(&_sbss);
     pDst = (uint8_t*)&_sbss;
     for (uint32_t i = 0; i<size; i++) {
         *pDst++ = 0;
